Refactor/fix AAS router
* Fix codestyle of legacy code.

* Compactify/realign many data structures.

* Reorganize code for lesser pointer chasing.

* Fix incorrect testing of "disable area" flags
for the next area (due to using reverse reachabilities
in routing the tested area was the current area).

* Fix incorrect legacy travel time estimations. The AAS
compiler does not assign "normal" "presence type" for areas
that have small height, even if there is a void area above.
The travel time was calculated for crouching travel type
greatly increasing travel time estimations.