Utilize SSE2 for fast area vis list scanning
While a sequential scan is already faster
than doing a ray-cast in a collision world
(that has been replaced by the scan)
there is an oppurtunity to speed it up
even more that should not be neglected.

This instruction set is almost 20 y/o and
should be supported on every target x86 CPU.
Assume it's always present if compiling for x86.