Add CM_TopNodeForBox() call and top node hints
An introduction of this allows to "scissor"
a map region for repeated collision calls
as long as they are not intended to leave
initial box bounds the node fully encloses.

AI and many other parts of the code could get
great benefits from using this.

This approach is not limited to BSP and is
useful for any kind of spatial hierarchy.

We can use a similar optimization for AAS
that has its own BSP tree and speed up
point area num sampling substantially.