S: Reuse "flood fill" results on the coarse path
Dijkstra algorithm builds a shortest path from a leaf
to every other leaf. While we still have to perform
point-to-point path problem solving for every
leaves pair on the fine path due to necessity to scale
edge weights along a built path to build alternatives,
alternatives are not used for the coarse propagation.
This allows to greatly reduce algorithmical complexity
of the coarse propagation path code.