AI: Add a better spot sort criteria separation
The old approach with a single score affected by
many criteria has never been working as it was
intented. It's better to keep all sub-scores
separate and use a lexicographical comparison
(a minor addition is that an inferior criterion
may affect a superior score if there's no complete
separation requested for the latter one).