AI,Sane: Drop Recast/Detour usage
This is incompatible with the primary AAS
navigation system and was only used for
getting a better tesselation of ground
areas for fallback kinds of movement.
Given the fact that the mesh is built upon
a buggy AAS data it was not really useful.
The lack of proper mesh debug visualisation
tools was not helpful either.

While the AAS considering its volumetric
nature is superior in every aspect Recast
could still be useful in future. The code
is preserved in the repo history.